pyproject.toml
[project]
name = "urlshortener"
version = "1.0.0"
description = "A full-featured URL shortener API built with FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.30,<3.0.0",
    "aiosqlite>=0.20.0,<1.0.0",
    "alembic>=1.13.0,<2.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "qrcode[pil]>=7.4,<9.0",
    "Pillow>=10.0.0,<13.0.0",
    "slowapi>=0.1.9,<1.0.0",
    "user-agents>=2.2.0,<3.0.0",
    "python-multipart>=0.0.9",
    "validators>=0.28.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<10.0.0",
    "pytest-asyncio>=0.23.0,<2.0.0",
    "httpx>=0.27.0,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
